+++ lib/src/certs/resign_committee_cold_cert.c
@@ -137,7 +137,7 @@ cardano_resign_committee_cold_cert_from_cbor(cardano_cbor_reader_t* reader, card
 
   static const char* validator_name = "resign_committee_cold_cert";
 
-  cardano_error_t expect_array_result = cardano_cbor_validate_array_of_n_elements(validator_name, reader, EMBEDDED_GROUP_SIZE);
+  cardano_error_t expect_array_result = cardano_cbor_validate_array_of_n_elements(validator_name, reader, (uint32_t)EMBEDDED_GROUP_SIZE);
 
   if (expect_array_result != CARDANO_SUCCESS)
   {
