+++ lib/src/proposal_procedures/committee.c
@@ -143,7 +143,7 @@ cardano_committee_from_cbor(cardano_cbor_reader_t* reader, cardano_committee_t**
 
   static const char* validator_name = "committee";
 
-  cardano_error_t expect_array_result = cardano_cbor_validate_array_of_n_elements(validator_name, reader, EMBEDDED_GROUP_SIZE);
+  cardano_error_t expect_array_result = cardano_cbor_validate_array_of_n_elements(validator_name, reader, (uint32_t)EMBEDDED_GROUP_SIZE);
 
   if (expect_array_result != CARDANO_SUCCESS)
   {
